move router builder to dedicated file
